getting the user from the db instead the token

+++ backend/src/module/user/user.service.ts
@@ -5,7 +5,6 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
 import * as bcrypt from 'bcryptjs';
-import { error } from 'console';
 
 @Injectable()
 export class UserService {
@@ -110,7 +109,15 @@ export class UserService {
 }
 
   async remove(id:number){
-  const user = await this.findOne(id);
+  const user = await this.userRepo
+    .createQueryBuilder('user')
+    .addSelect('user.password')
+    .where('user.id = :id', { id })
+    .getOne();
+
+  if (!user) {
+    throw new NotFoundException('User not found');
+  }
 
   await this.userRepo.remove(user);
